requirements.txt
Flask
google-generativeai
pypdf 